requirements.txt
# 手書きノート永久保存プロジェクト - 必要なPythonパッケージ

# 画像処理・PDF生成
Pillow>=10.0.0
img2pdf>=0.5.0

# OCR処理（将来の拡張用）
# tesseract-ocr  # システムレベルでインストールが必要
# pytesseract>=0.3.10

# その他のユーティリティ
pathlib2>=2.3.7  # Python 3.4+ では標準ライブラリに含まれているが、互換性のため
